findbugs: convert to double on integer values before casting them
Doing division on integers truncates the result to the integer value
closest to zero.

ICAST: integral division result cast to double or float (ICAST_IDIV_CAST_TO_DOUBLE)

+++ frontend/webadmin/modules/gwt-common/src/main/java/org/ovirt/engine/ui/common/widget/IntegerSlider.java
@@ -25,7 +25,7 @@ public class IntegerSlider extends SliderBar {
         super(minValue, maxValue, formatter, resources);
         this.setNumLabels(9);
         this.setNumTicks(9);
-        this.setStepSize((maxValue - minValue + 1) / 10);
+        this.setStepSize(((double) maxValue - (double) minValue + 1f) / 10f);
         this.setPixelSize(166, 30);
     }
